kadai14

+++ src/main/java/com/techacademy/KadaiFirstController.java
@@ -43,8 +43,8 @@ public class KadaiFirstController {
     }
     @GetMapping("/divide/{val1}/{val2}")
     public String calcDivide(@PathVariable int val1, @PathVariable int val2) {
-        int res = 0;
-        res = val1 / val2;
+        double res = 0.0;
+        res = (double) val1 / val2;
         return "実行結果：" + res;
     }
 }
